requirements.txt
fastapi
uvicorn
python-dotenv
supabase
requests
numpy
pydantic
pytest
